feat(search): 商品搜索

+++ mall-search/search-server/src/main/java/com/study/mall/vo/SearchVo.java
@@ -44,7 +44,7 @@ public class SearchVo {
     /**
      * 总页码
      */
-    private Integer totalPages;
+    private Long totalPages;
 
     @Data
     public static class BrandVo {
